Support connecting to networks with misconfigured subnet masks.
In K and earlier, we would connect to a network where the gateway
was not covered by the subnet mask of the IP address. This is an
invalid configuration, but it used to work, and other OSes appear
to accept it too, so support it.

Bug: 19067207

(cherry picked from commit 2d3ecdc5ac27a0fbd6721a9d436c95def0de3c66 [formerly 2dfb79a54adeb4bcf1f62332a9db467fce302ced])

Change-Id: I80088f291466dbd5a47f360dcc1620acee5cf57e

Former-commit-id: 8316b819b2837b6ad25957328ba7b2fdaf705fa9