Accessibility: Sometimes cannot interact with nav bar items.
If there is a window with the accessibility focus we want to click
on the accessibility focused view in this window. The logic to
compute the bounds of the window was using the wrong window id,
hence getting an incorrect result. As a consequence in some cases
the user could not click on accessiiblity focused controls in the
nav bar.

bug:18889611

Change-Id: I89aee3ae2ffe27fe29819049c287a7155154c65b

Former-commit-id: 10a053e474e2a566b49a1e4c9e2ec4c8cc0013bd