Refine resolution order for typeface and fontFamily
Previously any fontFamily value on a TextView would override a typeface
value, even if the fontFamily is from a TextAppearance (for example,
from the theme) and the typeface is explicitly set. This patch changes
the resolution order to fontFamily set directly on the TextView,
typeface set directly on the TextView, fontFamily from TextAppearance,
typeface from TextAppearance.

Bug: 16154223
Change-Id: I45c1e511fba8f64eb236200e3fa2e885c02b59dc

Former-commit-id: 42b3024efdec647266c1ae54a2a7c2c46915b380