Fix QS bluetooth tile accessibility issue
Bluetooth tile was not announcing its state when the top half was
clicked.  This is because if handleUpdateState was triggered on
the view before it became dual then it would not get all of the dual
accessibility information.  So if the dual state changes on a tile
make sure to call handleUpdateState so it can be handled appropriately.

Bug: 19155633
Change-Id: Ifd053c67d8ddd3230086517c9d479069556e8b56

Former-commit-id: 8ba572a3277f0c729a7fcce67bbf6e101667d5f9