Add frameworks classes and methods necessary for VVM syncadapter structure.
+ VvmSyncService is the base class for all visual voicemail sync
services that run the visual voicemail sync adapter. This class handles
writing to the voicemail provider and receiving changes from the
voicemail provider.
+ AuthenticatorService is a stub class for visual voicemail sync
adapters that do not use an Account (all sync adapters must have an
account associated with them, so a stub account is necessary).
+ Voicemail parcelable object to pass voicemail data around easier. This
is mostly copied from VoicemailProviderDemo
+ Extra helper methods in VoicemailContract to help insert and delete
from the voicemail provider.
+ Add multi-sim fields to VoicemailContract.Voicemails

Bug: 19236241
Change-Id: I603e3e5908704cd043e46221680d8bb600ed2cf4

Former-commit-id: 7c07dfa5529dd8049a526de227cae8abc2f72226