Apply stack dim layer to entire screen for system windows.
System windows like alert dialogs are attached to the currently
focused stack even when the stack isn't full screen. This causes
the dim not to apply to other visible stacks when a system
window is up. Change applies the dim to the entire screen for
system windows since they extend outside a resized stack.

Bug: 19332229
Change-Id: I94affa6a652326bf2cf14b1b8bf54862c111c540

Former-commit-id: 498e8c92d71ff4ccfd2a77bdc53160e8dc93c7fe