Fix issue where pruned UsageStats files would not be removed from index
This would cause an exception to be thrown when querying stats
that included a deleted file and cause only in-memory stats to be
returned.

This change now re-indexes after deleting files.

Furthermore, we continue reading UsageStats files in order
to return more useful data if some other issue (file corruption)
leads us to fail reading a file.

Change-Id: I4a52739624d68e719e3d7d324a0b16709a62ac7a

Former-commit-id: 3a0831bfe6e69a2ad11b9f6fae04fb27daa40a88