Fix crash in UndoManager after parceling TextView
UndoOwner maintains a manual reference count of the number of undo operations
it is associated with, but this count is not restored (and hence becomes zero)
when UndoManager is parceled and unparceled. This can cause the count to
underflow on subsequent text editing.

A test for this will land separately in CTS.

Bug: 19568283
Change-Id: Ic50890828db9679c7cef805388957d66dc75422d

Former-commit-id: f143ace3aa80227b7612c0a9428b2de996c4920a