Correct docs to specify leading slash in paths.
The guide documentation for the manifest data element contains a
grammar that implies leading slashes are not required for the path
part of a URI string. Also the path documentation does not mention
the need for a leading slash.

Grammar from documentation:
<scheme>://<host>:<port>/[<path>|<pathPrefix>|<pathPattern>]
                        ^
                        |
Included "/" -----------+

All internal usages of path, pathPrefix, and pathPattern, have a /
starting out the string and these attributes don't work without that
/.

This change updates the grammar and the documentation for the path
attributes.

Fixes bug 19540035

Change-Id: I931a25fd1c10c01eaa32a7cfa82cca256b129efb

Former-commit-id: a316f3eac3c9bd66098544d4d85aadaa0485f789