Indicate local and remote RX/TX capabilities for VT.
This provides more granular API support for representing the mix of
state and capabilities of local and remote video clients. For our
implementation, we are primarily concerned with bidirectional video
calling. But, this provides support for representing certain error
scenarios, and flexibility for other parties which may desire to
implement one-way video calling.

Change-Id: I0373ffd377c3547fa8dd98ab5fde4f8d616a1d2f

Former-commit-id: 5e9e8bb30e4d26eba46684c03ca950b602dfb830