Do not hide any call log entries when they are added.
When a call log entry is added, and it's phone account does not match a
currently registered one, we set it to hidden.  This code was built for
the calllog restore case where call log entries would be hidden when
added if the original phone account wasn't also
present on the new device (where the restore is being performed).

We no longer do that so we're removing the code that sets any call log
entry to hidden.

Change-Id: I26ee27369e94c73446f7553f84cd4d8d4f2ff658

Former-commit-id: 56027514f37e3184aeb7830237d753ad720df5ac