DO NOT MERGE Do not hide any call log entries when they are added.
When a call log entry is added, and it's phone account does not match a
currently registered one, we set it to hidden.  This code was built for
the calllog restore case where call log entries would be hidden when
added if the original phone account wasn't also
present on the new device (where the restore is being performed).

We no longer do that so we're removing the code that sets any call log
entry to hidden.

-- Resubmitting since this change was lost to a merge conflict --

Change-Id: I1ef094d5a35063e8f89cd1ecb1e5a0b59361781c

Former-commit-id: 14c9121e3c8f09c31f5de0d23c577263da665f02