Avoid rebuilding epoll set when PFD status report is done.
If we close the PFD status channel before returning, then the Looper
will take care of unregistering it correctly but it will end up
doing extra work to rebuild the epoll set to work around a limitation
of the epoll interface.  If we unregister the FD first then this
extra work will not be required so it's slightly more efficient
this way.

Bug: 19715279
Change-Id: I0f7b90f3458adf1894b7561c05fa842fdb87b498

Former-commit-id: b2a1985b139b6c8e10fb2cf87bfc4170dd7f2e3c