An updater for all timezone data on a device
The existing one, being deleted here, did not work properly: it only
updated the file used by libcore and bionic, it did not update the ICU
data.

Most of the installation logic exists in code in libcore/tzdata that is
independent of the server code so that it can be tested.

Bug: 19941636
Change-Id: Id0985f8c5be2f12858ee8bf52acf52bdb2df8741

Former-commit-id: 5a680f2e2738137f1566da57ac978b3ff33fb0da