Add voicemail builder method for updating/syncing.
The id of the voicemail is necessary to update a voicemail in the
database and the source data (or unique remote identifier for the
voicemail) is necessary for syncing remotely. Whenever performing a sync
either of data to the server or vise versa, create a voicemail object
using this new method.

Bug: 19236241
Change-Id: Ibc1fd8c1922be13605987d9a3dcd327c438951c8

Former-commit-id: d3e28b78a3107dd5eb00ac3f6130d9fcbc2116ad