Update progress_dialog on watch to use small spinner
The new material design theme causes the spinner to give us even
less room on the progress dialog screen.  This is causing cut off
text.

This change adds an alternate layout for the progress_dialog on
watches so that we can use a small spinner to make our text fit
better.

BUG: 18388692
BUG: 16493311
Change-Id: I2537aa2c1ee6329838c9c0c16f511c944fa86050

Former-commit-id: 457b8838cb7b9383bbad832ed35e8a58c7b6051b