Fixed issue with paused activity still freezing display.
An activity freezes the display when it is relaunched when
configuration changes. If the activity takes time to launch and
another activity is launched before it is done, the activity that
froze the display will be paused and might not have the chance to
unfreeze the display. This will put WindowManager in an odd state.
We now unfreeze the display when an activity is done pausing in case
it was previously freezing the display.

Bug: 19823482
Change-Id: If5538aea639e06d0b8621646bf6b2e12d325287a

Former-commit-id: 07927bffa72e35ded65b5eff2815f1337715c416