Recognizes the stylus scale gesture in ScaleGestureDetector
If a user drags with a stylus (or mouse) across the screen with
the stylus primary (or mouse secondary) button clicked, this will
initiate a scale gesture that functions similarly to the quick
scale gesture: drag below starting point zooms in, drag above
starting point zooms out.

Combines some of the logic for quick scale and button
scale due to similarities.

Bug: 19620246
Change-Id: Ic53499d07cf6cef48b467dbb41f83cea433febad

Former-commit-id: 847d17fcba7e0fab3093d05b5405554df91c08e2