Remove unused constants in android_text_StaticLayout.cpp
The logic for computing line breaking moved to Minikin, and so the
various Unicode characters with special meaning for line breaks are no
longer used in android_text_StaticLayout.cpp (which is now a fairly
thin wrapper that just does JNI glue). This patch deletes the
constants, which makes the warnings-as-errors setting of our current
clang build happy.

Change-Id: I1bb3596ca4f0c78a8e185af9000c0c366fb9b020

Former-commit-id: ee101a3c6627e7cf871f81583bb62f719485a2bd