Start searching the dimension configuration for atlas from the biggest value.
The old way will almost always miss the MAX_SIZE * MAX_SIZE case.
Most of time, it will be fine.
But if the threshold is just fall a bit short from MAX_SIZE * MAX_SIZE, then
it is likely that no generated configuration can be bigger than the threshold.

At the same time, a power of 2 size texture is more likely be used in the new
way. Better for memory alignment (or usage) and potential performance.

b/19966623

Change-Id: I4683fd5dea347158bc05d95cc7d777cc391b7fba

Former-commit-id: 072be09aafc258a8d9d8c4f1dbf4f800b3983434