Don't use context in measureText(char[])
The measureText(CharSequence) method calls into measureText(char[]),
using a temporary buffer, which may be longer than the CharSequence,
and may have additional characters. The existing implementation was
treating the entire char[] as potential context, but this is
incorrect.

This patch does the measurement on just the substring; measureText
doesn't process context. There are other methods for that, such as
getTextRunAdvances.

Bug: 19962931
Change-Id: Ia00cc004f9fa9d45aaf028683e78404d66c1388f

Former-commit-id: 80b485daa64d0ef2339c133f0d9cbc3c2c977c32