Update GpsMeasurements APIs to reflect changes in GPS HAL. b/19938206
All new introduced APIs are hidden.

GpsMeasurement#isPseudorangeRateCorrected() is need to properly identify different types
of data provided by different partners. Given the nature of each GPS engine, this change is
required to accomodate the data that can be provided by many partners.
GpsNavigationMessage#getStatus() is required to ensure that the Location process is aware
if the state of the messages, and can act depending on it.

Change-Id: Ie34b1e69bf26a1bde86931ce61168421429bf8f8

Former-commit-id: 5e72118f193258259912c0a51688bd4434c8896e