Persist nickname and flags for volumes.
StorageManager now offers to persist a nickname and user flags for
active volumes.  This metadata is kept around and spliced into
any future VolumeInfo when the same UUID is present.  Current user
flags indicate "initialized" and "snoozed" states to control how
notifications are shown.

Notify listeners when metadata changes, and kick public notification
after a volume is initialized.  Make unique PendingIntents when
multiple volumes are active.

Beginnings of plumbing to ask for missing volumes.

Offer explicit accessors for VolumeInfo and DiskInfo to give better
path to documentation and deprecation.

Bug: 19993667
Change-Id: I3d8b68be83f43ba992d21d51cad5b775776d681c

Former-commit-id: d95d3bfb2b28a4f21f3fdcd740160c9a61eb0363