DO NOT MERGE Move Phone.java APIs into InCallService
This change is one step in the right direction of getting rid of
Phone.java.

Phone.java was deemed superfluous and this change moves the methods
previously exposed in Phone.java into InCallService.  This was done by
having InCallService listen to Phone and pipe the calls as appropriate.
However, state still lives in Phone and we eventually want to move that
code into InCallService.  That will be done in a later CL.

Bug: 20160495
Change-Id: Id142431c253c1f24f260da42e8bedd1eb2ce448b

Former-commit-id: 7e24657d07868739754fd53fa7da97cb05e2e27d