Set REQUIRE_PASSWORD_TO_DECRYPT to false when insecure screen lock is enabled
Some users run into a problem when choosing a secure screen lock with
accessiblity enabled that they will not be warned the screen lock may
not be available when they unlock their devices. One cause is
REQUIRE_PASSWORD_TO_DECRYPT is not set back to false when users choose
to use an insecure lock screen. So when they decide to change back to a
secure screen lock, the warning message will not be shown.

Bug: 18799150
Change-Id: I52bdc93e7a5ce3c4f7715057c73b4de280b443f6

Former-commit-id: 95bbbdd097fb9ac410ca829df4be4e90992d2b48