Fix GcSnapshot drawInLayer when clipping is used.
GcSnapshot only allocates an image for tinting that is the size of the
current clipping area.  When drawing to that image, the wrong
coordinate system was used since now we are dealing with only the
clipping rect we need to translate -clipX, -clipY the drawing so it's
painted at 0,0 of the new BufferedImage.

Change-Id: Id8639d5ea9bef36ac07743df332161d6eadc6a81

Former-commit-id: 8ab069f3aed15c7e5de32ff52a3b4710218f9895