Add annotation to Rect.intersect()
Failing to check the return result from intersect() can cause artifacts
because the developer may assume that the rect has been changed to the intersection.
In particular, when the two rects do not overlap, there is no change made to the
source rect. Instead, the method simply returns false. When developers do not
check that return value, they may use the source rect and get undefined results.

This change adds the @CheckResults annotation that will cause a Lint warning
when developers call this method without checking the return value.

Change-Id: I476367d74e712038c248c2379fb124734298fcc1

Former-commit-id: 5b3edbde20b446ecbfebe5d6ad0dcd735473ac75